requirements.txt
flask
Flask-Caching
python-nmap
